+++ front/src/components/signupform/SignUpForm.tsx
@@ -1,5 +1,6 @@
 import { ChangeEvent, MouseEvent, useEffect, useState } from "react";
-import { SignUpFormBlock, InputGroup, Button } from "./SignUpForm.element";
+import { SignUpFormBlock, InputGroup } from "./SignUpForm.element";
+import { Button } from "../common/Button";
 import { checkEmail, checkPassword } from "../../utils/validator";
 import { useNavigate } from "react-router-dom";
 import openEye from "../../assets/openeye.svg";
@@ -45,21 +46,16 @@ export const SignUpForm = () => {
   //패스워드 확인
   const [passwordConfirm, setPasswordConfirm] = useState("");
   const [isPasswordConfirmValid, setIsPasswordConfirmValid] = useState(false);
-  const [isPasswordConfirmVisible, setIsPasswordConfirmVisible] =
-    useState(false);
+  const [isPasswordConfirmVisible, setIsPasswordConfirmVisible] = useState(false);
 
   const handlePasswordConfirmChange = (e: ChangeEvent<HTMLInputElement>) => {
     const newPasswordConfirm = e.target.value;
     setPasswordConfirm(newPasswordConfirm);
     updatePasswordConfirmValid(password, newPasswordConfirm);
   };
-  const updatePasswordConfirmValid = (
-    _password: string,
-    _passwordConfirm: string
-  ) => {
+  const updatePasswordConfirmValid = (_password: string, _passwordConfirm: string) => {
     if (_password !== _passwordConfirm) setIsPasswordConfirmValid(false);
-    if (_password === _passwordConfirm || !_passwordConfirm)
-      setIsPasswordConfirmValid(true);
+    if (_password === _passwordConfirm || !_passwordConfirm) setIsPasswordConfirmValid(true);
   };
 
   const handlePasswordConfirmVisibleClick = () => {
@@ -105,14 +101,8 @@ export const SignUpForm = () => {
     <SignUpFormBlock>
       <InputGroup>
         <label htmlFor="email">이메일</label>
-        <input
-          name="email"
-          placeholder="ssafe11@gmail.com"
-          onChange={handleEmailChange}
-        />
-        <span>
-          {email && !isEmailValid && "유효하지 않은 이메일 형식입니다."}
-        </span>
+        <input name="email" placeholder="ssafe11@gmail.com" onChange={handleEmailChange} />
+        <span>{email && !isEmailValid && "유효하지 않은 이메일 형식입니다."}</span>
       </InputGroup>
       <InputGroup>
         <label htmlFor="password">비밀번호</label>
@@ -122,14 +112,8 @@ export const SignUpForm = () => {
           type={isPasswordVisible ? "text" : "password"}
           onChange={handlePasswordChange}
         />
-        <img
-          src={isPasswordVisible ? openEye : closeEye}
-          onClick={handlePasswordVisibleClick}
-          alt=""
-        />
-        <span>
-          {password && !isPasswordValid && "유효하지 않은 비밀번호 형식입니다."}
-        </span>
+        <img src={isPasswordVisible ? openEye : closeEye} onClick={handlePasswordVisibleClick} alt="" />
+        <span>{password && !isPasswordValid && "유효하지 않은 비밀번호 형식입니다."}</span>
       </InputGroup>
       <InputGroup>
         <label htmlFor="passwordConfirm">비밀번호 확인</label>
@@ -139,16 +123,8 @@ export const SignUpForm = () => {
           type={isPasswordConfirmVisible ? "text" : "password"}
           onChange={handlePasswordConfirmChange}
         />
-        <img
-          src={isPasswordConfirmVisible ? openEye : closeEye}
-          onClick={handlePasswordConfirmVisibleClick}
-          alt=""
-        />
-        <span>
-          {passwordConfirm &&
-            !isPasswordConfirmValid &&
-            "비밀번호가 일치하지 않습니다."}
-        </span>
+        <img src={isPasswordConfirmVisible ? openEye : closeEye} onClick={handlePasswordConfirmVisibleClick} alt="" />
+        <span>{passwordConfirm && !isPasswordConfirmValid && "비밀번호가 일치하지 않습니다."}</span>
       </InputGroup>
       <Button active={isActive} onClick={handleBtnClick}>
         회원가입
